change type in my_strlen

+++ string.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-size_t my_strlen(const char *s)
+int my_strlen(const char *s)
 {
    const char *p = s;
